@@ -1,6 +1,6 @@
 {
   "name": "@jsenv/node-module-import-map",
-  "version": "11.1.2",
+  "version": "12.0.0-alpha.0",
   "description": "Generate importmap for node_modules.",
   "license": "MIT",
   "repository": {
@@ -37,32 +37,29 @@
     "upload-coverage": "node --no-warnings ./script/upload-coverage/upload-coverage.js",
     "generate-commonjs-bundle": "node --no-warnings ./script/generate-commonjs-bundle/generate-commonjs-bundle.js",
     "generate-import-map": "node --no-warnings ./script/generate-import-map/generate-import-map.js",
-    "install-git-hooks": "node --no-warnings ./script/install-git-hooks/install-git-hooks.js",
     "dist": "npm run generate-commonjs-bundle",
-    "postinstall": "npm run generate-import-map && npm run install-git-hooks",
+    "postinstall": "npm run generate-import-map",
     "prepublishOnly": "node ./script/transform-package/remove-postinstall.js && npm run dist",
-    "postpublish": "node ./script/transform-package/restore-postinstall.js",
-    "git-hook-pre-commit": "npm run prettier-format-stage"
+    "postpublish": "node ./script/transform-package/restore-postinstall.js"
   },
   "dependencies": {
-    "@jsenv/cancellation": "1.4.1",
-    "@jsenv/import-map": "6.7.3",
+    "@jsenv/cancellation": "1.4.2",
+    "@jsenv/import-map": "6.7.4",
     "@jsenv/logger": "3.3.1",
-    "@jsenv/util": "3.5.2"
+    "@jsenv/util": "3.6.3"
   },
   "devDependencies": {
-    "@jsenv/assert": "2.1.0",
-    "@jsenv/codecov-upload": "3.4.0",
-    "@jsenv/core": "12.2.2",
-    "@jsenv/eslint-config": "12.3.1",
-    "@jsenv/git-hooks": "1.3.3",
-    "@jsenv/node-module-import-map": "11.0.1",
-    "@jsenv/prettier-check-project": "5.5.0",
+    "@jsenv/assert": "2.2.1",
+    "@jsenv/codecov-upload": "3.4.1",
+    "@jsenv/core": "14.0.0-alpha.10",
+    "@jsenv/eslint-config": "12.4.1",
+    "@jsenv/node-module-import-map": "11.1.2",
+    "@jsenv/prettier-check-project": "5.6.0",
     "@jsenv/github-release-package": "1.2.2",
     "@jsenv/package-publish": "1.5.1",
     "@jsenv/prettier-config": "1.2.0",
     "babel-eslint": "11.0.0-beta.0",
-    "eslint": "6.8.0",
-    "prettier": "2.0.2"
+    "eslint": "7.4.0",
+    "prettier": "2.0.5"
   }
 }
