@@ -1,8 +1,18 @@
-import { generateImportMapForProjectPackage } from "@jsenv/node-module-import-map"
-import * as jsenvConfig from "../../jsenv.config.js"
+import {
+  getImportMapFromNodeModules,
+  generateImportMapForProject,
+} from "@jsenv/node-module-import-map"
+import { projectDirectoryUrl, importMapFileRelativeUrl } from "../../jsenv.config.js"
 
-generateImportMapForProjectPackage({
-  ...jsenvConfig,
-  importMapFile: true,
-  jsConfigFile: true,
-})
+generateImportMapForProject(
+  [
+    getImportMapFromNodeModules({
+      projectDirectoryUrl,
+    }),
+  ],
+  {
+    projectDirectoryUrl,
+    importMapFileRelativeUrl,
+    jsConfigFile: true,
+  },
+)
